Промежуточный фикс WangLandauParallel

+++ wanglandau.cpp
@@ -48,7 +48,7 @@ vector<double> WangLandau::dos(PartArray &sys,const int intervals,const int step
         //повторяем алгоритм сколько-то шагов
         for (int i=0;i<steps;i++){
             double rand = (double)config::Instance()->rand()/(double)config::Instance()->rand_max;
-            int partNum = (int)floor(rand*count);
+            int partNum = (int)floor(rand*(double)count);
             sys.parts[partNum]->rotate();
             eNew = sys.calcEnergy1FastIncremental(eInit);
